+++ Database/Updates/Shard/2019-01-19-01-Emote_Table_Display_Change.sql
@@ -1,0 +1,4 @@
+USE `ace_shard`;
+
+ALTER TABLE `biota_properties_emote_action` 
+CHANGE COLUMN `display` `display` BIT(1) NULL DEFAULT NULL ;