+++ Database/Updates/World/2019-01-16-01-Recipe_Mods_Table_Changes.sql
@@ -1,0 +1,19 @@
+USE `ace_world`;
+
+ALTER TABLE `recipe_mods_bool` 
+CHANGE COLUMN `unknown_1` `source` INT(10) NOT NULL ;
+
+ALTER TABLE `recipe_mods_d_i_d` 
+CHANGE COLUMN `unknown_1` `source` INT(10) NOT NULL ;
+
+ALTER TABLE `recipe_mods_float` 
+CHANGE COLUMN `unknown_1` `source` INT(10) NOT NULL ;
+
+ALTER TABLE `recipe_mods_i_i_d` 
+CHANGE COLUMN `unknown_1` `source` INT(10) NOT NULL ;
+
+ALTER TABLE `recipe_mods_int` 
+CHANGE COLUMN `unknown_1` `source` INT(10) NOT NULL ;
+
+ALTER TABLE `recipe_mods_string` 
+CHANGE COLUMN `unknown_1` `source` INT(10) NOT NULL ;