+++ Source/ACE.Database/Models/World/RecipeModsInt.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 
 namespace ACE.Database.Models.World
@@ -10,7 +10,7 @@ namespace ACE.Database.Models.World
         public int Stat { get; set; }
         public int Value { get; set; }
         public int Enum { get; set; }
-        public int Unknown1 { get; set; }
+        public int Source { get; set; }
 
         public RecipeMod RecipeMod { get; set; }
     }